+++ src/Northwind.AB.Sales.Backend.BusinessObjects/ValueObjects/OrderDetail.cs
@@ -2,4 +2,4 @@
 public record struct OrderDetail(
     int ProductId,
     decimal UnitPrice,
-    short Quantity);
+    int Quantity);
